requirements.txt
langchain
streamlit
sqlalchemy
alembic
python-dotenv
psycopg2-binary
langchain_community
openai
langchain-openai
streamlit-chatbox
fastapi
bcrypt
uvicorn
PyJWT